move boosts logic to HomePage
