+++ frontend/src/pages/HomePage.jsx
@@ -1,12 +1,17 @@
 import {
+  addBoost,
+  removeBoost,
+  updatedBoostCount,
+} from "../firebase/functions/boostFunctionality";
+import CreatePost from "../components/CreatePost/CreatePost";
+import PortfolioCard from "../components/PortfolioCard/PortfolioCard";
+import { Card } from "../components/ui/index";
+import {
   fetchPortfolio,
   fetchRoleById,
   fetchUserById,
 } from "../firebase/functions/index";
-import CreatePost from "../components/CreatePost/CreatePost";
-import PortfolioCard from "../components/PortfolioCard/PortfolioCard";
 import { useEffect, useState } from "react";
-import { Card } from "../components/ui/index";
 
 function HomePage({ currentUser }) {
   const [portfolios, setPortfolios] = useState([]);
@@ -54,6 +59,27 @@ function HomePage({ currentUser }) {
     getData();
   }, []);
 
+  const handleBoostClick = async (portfolioId) => {
+    try {
+      const portfolioItem = portfolios.find((p) => p.id === portfolioId);
+      const currentBoostCount = portfolioItem?.boostCount || 0;
+      const newBoostCount = currentBoostCount + 1;
+
+      await addBoost(portfolioId);
+      await updatedBoostCount(portfolioId, newBoostCount);
+
+      const updatedPortfolios = portfolios.map((item) =>
+        item.id === portfolioId
+          ? { ...item, boostCount: newBoostCount, boosted: !item.boosted }
+          : item
+      );
+
+      setPortfolios([...updatedPortfolios]);
+    } catch (error) {
+      console.error("Error handling boost click:", error);
+    }
+  };
+
   return (
     <div className="grid grid-cols-1 gap-[3.13rem] justify-items-center">
       <CreatePost />
@@ -68,7 +94,7 @@ function HomePage({ currentUser }) {
                 portfolio={portfolio}
                 user={users[portfolio.userId]}
                 role={roles[portfolio.userId?.roleId]}
-                // handleBoostClick={handleBoostClick}
+                handleBoostClick={handleBoostClick}
               />
             ))}
           </Card>
